xml/parsing/manakai/eldecls-1.dat

#data
<!DOCTYPE a [
<!ELEM
#errors
1;1;no XML decl;;;s
2;7;bogus comment;;;m
2;7;parser:EOF;;;m
2;7;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a [
<!ELEMENT
#errors
1;1;no XML decl;;;s
2;10;parser:EOF;;;m
2;10;parser:EOF;;;m
2;10;no root element;;;m
#document
| <!DOCTYPE a>

#data escpaed
<!DOCTYPE a [
<!ELEMENT>
]><a/>
#errors
1;1;no XML decl;;;s
2;10;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data escpaed
<!DOCTYPE a [
<!ELEMENT foo(a)>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;no space between params;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo (a)>
| <{} a>

#data escpaed
<!DOCTYPE a [
<!ELEMENT foo>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data escpaed
<!DOCTYPE a [
<!ELEMENT\u0020
#errors
1;1;no XML decl;;;s
2;10;no space between params;;;m
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a [
<!ELEMENT \u0020
#errors
1;1;no XML decl;;;s
2;12;parser:EOF;;;m
2;12;parser:EOF;;;m
2;12;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a [
<!ELEMENT abc
#errors
1;1;no XML decl;;;s
2;14;parser:EOF;;;m
2;14;parser:EOF;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a [
<!ELEMENT abc\u0020
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b EMPTY>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b EMPTY>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b amny>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:unknown content keyword;;amny;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b em>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:unknown content keyword;;em;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY*>
]><a/>
#errors
1;1;no XML decl;;;s
2;1;xml:dtd:unknown content keyword;;ANY*;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (
#errors
1;1;no XML decl;;;s
2;14;parser:EOF;;;m
2;14;parser:EOF;;;m
2;14;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (\u0009
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ()>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b ()>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ()+>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b ()>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ()*>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b ()>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ()?>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b ()>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x\u0020
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x,
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x|
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x&
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x,\u0020
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x|\u0020
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x,y
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x,y\u0020
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x,y ) >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x, y)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ( zzz |  rrr )>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (zzz | rrr)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)y>
]><a/>
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x))>
]><a/>
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x) )>
]><a/>
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x)\u0009
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)+
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)*
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)?
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (x)?\u0020
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)+y>
]><a/>
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)+>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x#
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x #
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b ((\u0020
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((x
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b ((x\u0020
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((x?
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((x)
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((x)?
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b ((x)?\u0020
#errors
1;1;no XML decl;;;s
2;19;parser:EOF;;;m
2;19;parser:EOF;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b ((x)\u0020
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b ((x))>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b ((x))>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (   (x)  )>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b ((x))>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x+)>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x+)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)  >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x+)  >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x+)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)*  >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)*>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)?   >
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)?>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x+   )>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x+)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x  +)>
]><a/>
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)   +   >
]><a/>
#errors
1;1;no XML decl;;;s
2;19;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)  *>
]><a/>
#errors
1;1;no XML decl;;;s
2;18;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)  ?>
]><a/>
#errors
1;1;no XML decl;;;s
2;18;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x?,x*)>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x?, x*)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x, (y)*)>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x, (y)*)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (#
#errors
1;1;no XML decl;;;s
2;15;parser:EOF;;;m
2;15;parser:EOF;;;m
2;15;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (#PCD
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (#PCDATA
#errors
1;1;no XML decl;;;s
2;21;parser:EOF;;;m
2;21;parser:EOF;;;m
2;21;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (#PCDATA\u0020
#errors
1;1;no XML decl;;;s
2;22;parser:EOF;;;m
2;22;parser:EOF;;;m
2;22;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (#PCDATA | zz | aa)>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (#PCDATA | zz | aa)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (#PCDATA | #PCDATA , a)>
]><a/>
#errors
1;1;no XML decl;;;s
#XXX
#document
| <!DOCTYPE a>
|   <!ELEMENT b (#PCDATA | #PCDATA | a)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (#PCDATA)*>
]><a/>
#errors
1;1;no XML decl;;;s
#document
| <!DOCTYPE a>
|   <!ELEMENT b (#PCDATA)*>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a,b>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a  >
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a ,>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((ab+>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a,(>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a) >
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a))>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ((((a)+>
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (((a)+ >
]><a/>
#errors
1;1;no XML decl;;;s
2;20;unclosed cmgroup;;;m
#document
| <!DOCTYPE a>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x y)*>
]><a/>
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b (x)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b (x#
#errors
1;1;no XML decl;;;s
2;16;parser:EOF;;;m
2;16;parser:EOF;;;m
2;16;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x #
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x,#
#errors
1;1;no XML decl;;;s
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x, #
#errors
1;1;no XML decl;;;s
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)#
#errors
1;1;no XML decl;;;s
2;16;xml:dtd:no group delimiter;;;m
2;17;parser:EOF;;;m
2;17;parser:EOF;;;m
2;17;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x) #
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:no group delimiter;;;m
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)+#
#errors
1;1;no XML decl;;;s
2;17;xml:dtd:no group delimiter;;;m
2;18;parser:EOF;;;m
2;18;parser:EOF;;;m
2;18;no root element;;;m
#document
| <!DOCTYPE a>

#data
<!DOCTYPE a[
<!ELEMENT b (x)? #
#errors
1;1;no XML decl;;;s
2;18;xml:dtd:no group delimiter;;;m
2;19;parser:EOF;;;m
2;19;parser:EOF;;;m
2;19;no root element;;;m
#document
| <!DOCTYPE a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b \u0000a>
]><a/>
#errors escaped
1;1;no XML decl;;;s
2;1;xml:dtd:unknown content keyword;;\u0000a;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b x\u0000a>
]><a/>
#errors escaped
1;1;no XML decl;;;s
2;1;xml:dtd:unknown content keyword;;x\u0000a;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b >
| <{} a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (\u0000a)>
]><a/>
#errors
1;1;no XML decl;;;s
2;14;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b (\uFFFDa)>
| <{} a>

#data escaped
<!DOCTYPE a[
<!ELEMENT b (a\u0000a)>
]><a/>
#errors
1;1;no XML decl;;;s
2;15;NULL;;;m
#document escaped
| <!DOCTYPE a>
|   <!ELEMENT b (a\uFFFDa)>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b EMPTY>
<!ELEMENT b ANY>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;duplicate element decl;;b;m
#document
| <!DOCTYPE a>
|   <!ELEMENT b EMPTY>
| <{} a>

#data
<!DOCTYPE a[
<!ELEMENT b ANY>
%x;
<!ELEMENT c ANY>
]><a/>
#errors
1;1;no XML decl;;;s
3;1;entity not declared;;%x;;m
3;1;stop processing;;;i
#document
| <!DOCTYPE a>
|   <!ELEMENT b ANY>
|   <!ELEMENT c ANY>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (a)(>]><a/>
#errors
1;51;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo (a)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (a()>]><a/>
#errors
1;50;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo (a)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (a ()>]><a/>
#errors
1;51;xml:dtd:no group delimiter;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo (a)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (a)|>]><a/>
#errors
1;51;string after md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo (a)>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (*)>]><a/>
#errors
1;49;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo ()>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (+)>]><a/>
#errors
1;49;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo ()>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (,)>]><a/>
#errors
1;49;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo ()>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (?)>]><a/>
#errors
1;49;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo ()>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo (|)>]><a/>
#errors
1;49;xml:dtd:no group item;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo ()>
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo )>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo *>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo +>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo ?>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo ,>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo |>]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT foo >]><a/>
#errors
1;48;no md def;;;m
#document
| <!DOCTYPE a>
|   <!ELEMENT foo >
| <{} a>

#data
<?xml version="1.0"?><!DOCTYPE a[<!ELEMENT >]><a/>
#errors
1;44;no md name;;;m
#document
| <!DOCTYPE a>
| <{} a>
